+++ app/Templates/Mail/Message.php
@@ -12,6 +12,7 @@ use Nextras\Orm\Relationships\OneHasMany;
 final class Message {
 	public function __construct(
 		public ?string $accountNumber,
+		public ?string $accountNumberIban,
 		public string $eventName,
 		public string $eventNameShort,
 		public string $dateFormat,